0.2.1: OpenCastor embodiment reads canvas size per reference; deny sub-codes; lint clean
calibrate.py --reference was accepted and ignored, so teaching corners for a
custom canvas fitted against the wrong sheet; corner_canvas_mm() now resolves
from the spec. Gateway denies surface detail.telemetry.deny (unsafe_pose names
the arm.reach_point remedy). docs/opencastor.md: 150 x 200 mm sheet, angled
pen mount, Bob's working -E flags. Tests derive expectations from the spec.

+++ src/sacpaint/scorers.py
@@ -129,7 +129,7 @@ def _aligned_presence(
     c_ref, c_can = _centroid(r_box), _centroid(c_box)
     if c_ref is None or c_can is None:
         return 0.0, 0.0, 0.0, c_ref, c_can
-    dx, dy = int(round(c_ref[0] - c_can[0])), int(round(c_ref[1] - c_can[1]))
+    dx, dy = round(c_ref[0] - c_can[0]), round(c_ref[1] - c_can[1])
     shifted = np.zeros_like(c_crop)
     h, w = c_crop.shape
     src_y = slice(max(0, -dy), min(h, h - dy))
